+++ instrumentation/github.com/emicklei/go-restful/otelrestful/restful_test.go
@@ -25,12 +25,11 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"go.opentelemetry.io/contrib/instrumentation/github.com/emicklei/go-restful/otelrestful"
-
-	mocktrace "go.opentelemetry.io/contrib/internal/trace"
 	b3prop "go.opentelemetry.io/contrib/propagators/b3"
 	"go.opentelemetry.io/otel"
 	otelglobal "go.opentelemetry.io/otel/api/global"
 	oteltrace "go.opentelemetry.io/otel/api/trace"
+	"go.opentelemetry.io/otel/api/trace/tracetest"
 	otelkv "go.opentelemetry.io/otel/label"
 	"go.opentelemetry.io/otel/propagators"
 )
@@ -38,14 +37,14 @@ import (
 const tracerName = "go.opentelemetry.io/contrib/instrumentation/github.com/emicklei/go-restful/otelrestful"
 
 func TestChildSpanFromGlobalTracer(t *testing.T) {
-	otelglobal.SetTracerProvider(&mocktrace.TracerProvider{})
+	otelglobal.SetTracerProvider(tracetest.NewTracerProvider())
 
 	handlerFunc := func(req *restful.Request, resp *restful.Response) {
 		span := oteltrace.SpanFromContext(req.Request.Context())
-		_, ok := span.(*mocktrace.Span)
+		_, ok := span.(*tracetest.Span)
 		assert.True(t, ok)
 		spanTracer := span.Tracer()
-		mockTracer, ok := spanTracer.(*mocktrace.Tracer)
+		mockTracer, ok := spanTracer.(*tracetest.Tracer)
 		require.True(t, ok)
 		assert.Equal(t, tracerName, mockTracer.Name)
 		resp.WriteHeader(http.StatusOK)
@@ -65,14 +64,14 @@ func TestChildSpanFromGlobalTracer(t *testing.T) {
 }
 
 func TestChildSpanFromCustomTracer(t *testing.T) {
-	provider, _ := mocktrace.NewTracerProviderAndTracer(tracerName)
+	provider := tracetest.NewTracerProvider()
 
 	handlerFunc := func(req *restful.Request, resp *restful.Response) {
 		span := oteltrace.SpanFromContext(req.Request.Context())
-		_, ok := span.(*mocktrace.Span)
+		_, ok := span.(*tracetest.Span)
 		assert.True(t, ok)
 		spanTracer := span.Tracer()
-		mockTracer, ok := spanTracer.(*mocktrace.Tracer)
+		mockTracer, ok := spanTracer.(*tracetest.Tracer)
 		require.True(t, ok)
 		assert.Equal(t, tracerName, mockTracer.Name)
 		resp.WriteHeader(http.StatusOK)
@@ -91,7 +90,8 @@ func TestChildSpanFromCustomTracer(t *testing.T) {
 }
 
 func TestChildSpanNames(t *testing.T) {
-	provider, tracer := mocktrace.NewTracerProviderAndTracer(tracerName)
+	sr := new(tracetest.StandardSpanRecorder)
+	provider := tracetest.NewTracerProvider(tracetest.WithSpanRecorder(sr))
 
 	handlerFunc := func(req *restful.Request, resp *restful.Response) {
 		resp.WriteHeader(http.StatusOK)
@@ -111,30 +111,30 @@ func TestChildSpanNames(t *testing.T) {
 	w := httptest.NewRecorder()
 
 	container.ServeHTTP(w, r)
-	spans := tracer.EndedSpans()
+	spans := sr.Completed()
 	require.Len(t, spans, 1)
 	span := spans[0]
-	assert.Equal(t, "/user/{id:[0-9]+}", span.Name)
-	assert.Equal(t, oteltrace.SpanKindServer, span.Kind)
-	assert.Equal(t, otelkv.StringValue("foobar"), span.Attributes["http.server_name"])
-	assert.Equal(t, otelkv.IntValue(http.StatusOK), span.Attributes["http.status_code"])
-	assert.Equal(t, otelkv.StringValue("GET"), span.Attributes["http.method"])
-	assert.Equal(t, otelkv.StringValue("/user/123"), span.Attributes["http.target"])
-	assert.Equal(t, otelkv.StringValue("/user/{id:[0-9]+}"), span.Attributes["http.route"])
+	assert.Equal(t, "/user/{id:[0-9]+}", span.Name())
+	assert.Equal(t, oteltrace.SpanKindServer, span.SpanKind())
+	assert.Equal(t, otelkv.StringValue("foobar"), span.Attributes()["http.server_name"])
+	assert.Equal(t, otelkv.IntValue(http.StatusOK), span.Attributes()["http.status_code"])
+	assert.Equal(t, otelkv.StringValue("GET"), span.Attributes()["http.method"])
+	assert.Equal(t, otelkv.StringValue("/user/123"), span.Attributes()["http.target"])
+	assert.Equal(t, otelkv.StringValue("/user/{id:[0-9]+}"), span.Attributes()["http.route"])
 
 	r = httptest.NewRequest("GET", "/book/foo", nil)
 	w = httptest.NewRecorder()
 	container.ServeHTTP(w, r)
-	spans = tracer.EndedSpans()
-	require.Len(t, spans, 1)
-	span = spans[0]
-	assert.Equal(t, "/book/{title}", span.Name)
-	assert.Equal(t, oteltrace.SpanKindServer, span.Kind)
-	assert.Equal(t, otelkv.StringValue("foobar"), span.Attributes["http.server_name"])
-	assert.Equal(t, otelkv.IntValue(http.StatusOK), span.Attributes["http.status_code"])
-	assert.Equal(t, otelkv.StringValue("GET"), span.Attributes["http.method"])
-	assert.Equal(t, otelkv.StringValue("/book/foo"), span.Attributes["http.target"])
-	assert.Equal(t, otelkv.StringValue("/book/{title}"), span.Attributes["http.route"])
+	spans = sr.Completed()
+	require.Len(t, spans, 2)
+	span = spans[1]
+	assert.Equal(t, "/book/{title}", span.Name())
+	assert.Equal(t, oteltrace.SpanKindServer, span.SpanKind())
+	assert.Equal(t, otelkv.StringValue("foobar"), span.Attributes()["http.server_name"])
+	assert.Equal(t, otelkv.IntValue(http.StatusOK), span.Attributes()["http.status_code"])
+	assert.Equal(t, otelkv.StringValue("GET"), span.Attributes()["http.method"])
+	assert.Equal(t, otelkv.StringValue("/book/foo"), span.Attributes()["http.target"])
+	assert.Equal(t, otelkv.StringValue("/book/{title}"), span.Attributes()["http.route"])
 }
 
 func TestGetSpanNotInstrumented(t *testing.T) {
@@ -156,21 +156,21 @@ func TestGetSpanNotInstrumented(t *testing.T) {
 }
 
 func TestPropagationWithGlobalPropagators(t *testing.T) {
-	provider, tracer := mocktrace.NewTracerProviderAndTracer(tracerName)
+	provider := tracetest.NewTracerProvider()
 	otelglobal.SetTextMapPropagator(propagators.TraceContext{})
 
 	r := httptest.NewRequest("GET", "/user/123", nil)
 	w := httptest.NewRecorder()
 
-	ctx, pspan := tracer.Start(context.Background(), "test")
+	ctx, pspan := provider.Tracer(tracerName).Start(context.Background(), "test")
 	otelglobal.TextMapPropagator().Inject(ctx, r.Header)
 
 	handlerFunc := func(req *restful.Request, resp *restful.Response) {
 		span := oteltrace.SpanFromContext(req.Request.Context())
-		mspan, ok := span.(*mocktrace.Span)
+		mspan, ok := span.(*tracetest.Span)
 		require.True(t, ok)
 		assert.Equal(t, pspan.SpanContext().TraceID, mspan.SpanContext().TraceID)
-		assert.Equal(t, pspan.SpanContext().SpanID, mspan.ParentSpanID)
+		assert.Equal(t, pspan.SpanContext().SpanID, mspan.ParentSpanID())
 		w.WriteHeader(http.StatusOK)
 	}
 	ws := &restful.WebService{}
@@ -185,21 +185,21 @@ func TestPropagationWithGlobalPropagators(t *testing.T) {
 }
 
 func TestPropagationWithCustomPropagators(t *testing.T) {
-	provider, tracer := mocktrace.NewTracerProviderAndTracer(tracerName)
+	provider := tracetest.NewTracerProvider()
 	b3 := b3prop.B3{}
 
 	r := httptest.NewRequest("GET", "/user/123", nil)
 	w := httptest.NewRecorder()
 
-	ctx, pspan := tracer.Start(context.Background(), "test")
+	ctx, pspan := provider.Tracer(tracerName).Start(context.Background(), "test")
 	b3.Inject(ctx, r.Header)
 
 	handlerFunc := func(req *restful.Request, resp *restful.Response) {
 		span := oteltrace.SpanFromContext(req.Request.Context())
-		mspan, ok := span.(*mocktrace.Span)
+		mspan, ok := span.(*tracetest.Span)
 		require.True(t, ok)
 		assert.Equal(t, pspan.SpanContext().TraceID, mspan.SpanContext().TraceID)
-		assert.Equal(t, pspan.SpanContext().SpanID, mspan.ParentSpanID)
+		assert.Equal(t, pspan.SpanContext().SpanID, mspan.ParentSpanID())
 		w.WriteHeader(http.StatusOK)
 	}
 	ws := &restful.WebService{}
@@ -215,15 +215,16 @@ func TestPropagationWithCustomPropagators(t *testing.T) {
 }
 
 func TestMultiFilters(t *testing.T) {
-	provider, _ := mocktrace.NewTracerProviderAndTracer(tracerName)
+	sr := new(tracetest.StandardSpanRecorder)
+	provider := tracetest.NewTracerProvider(tracetest.WithSpanRecorder(sr))
 
 	wrappedFunc := func(tracerName string) restful.RouteFunction {
 		return func(req *restful.Request, resp *restful.Response) {
 			span := oteltrace.SpanFromContext(req.Request.Context())
-			_, ok := span.(*mocktrace.Span)
+			_, ok := span.(*tracetest.Span)
 			assert.True(t, ok)
 			spanTracer := span.Tracer()
-			mockTracer, ok := spanTracer.(*mocktrace.Tracer)
+			mockTracer, ok := spanTracer.(*tracetest.Tracer)
 			require.True(t, ok)
 			assert.Equal(t, tracerName, mockTracer.Name)
 			resp.WriteHeader(http.StatusOK)