+++ CHANGELOG.md
@@ -16,6 +16,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 - Add semantic version to `Tracer` / `Meter` created by instrumentation packages `otelsaram`, `otelrestful`, `otelmongo`, `otelhttp` and `otelhttptrace`. (#412)
 - Update instrumentation guidelines about tracer / meter semantic version. (#412)
+- Replace internal tracer and meter helpers by helpers from `go.opentelemetry.io/otel`. (#414)
 - gRPC instrumentation sets span attribute `rpc.grpc.status_code`. (#453)
 
 ## Fixed